requirements.txt
rsa==4.8
PyQt5==5.15.6
watchdog==2.1.7
pycryptodomex==3.9.0